Move model file
